+++ flake8_comprehensions.py
@@ -18,9 +18,7 @@ class ComprehensionChecker(object):
     def __init__(self, tree, *args, **kwargs):
         self.tree = tree
 
-    message_C100 = 'C100 Unnecessary generator - rewrite as a list comprehension.'
-    message_C101 = 'C101 Unnecessary generator - rewrite as a set comprehension.'
-    message_C102 = 'C102 Unnecessary generator - rewrite as a dict comprehension.'
+    message_C400 = 'C400 Unnecessary generator - rewrite as a {type} comprehension.'
 
     def run(self):
         for node in ast.walk(self.tree):
@@ -29,9 +27,10 @@ class ComprehensionChecker(object):
                 len(node.args) == 1 and
                 isinstance(node.args[0], ast.GeneratorExp)
             ):
-                if node.func.id == 'list':
-                    yield (node.lineno, node.col_offset, self.message_C100, type(self))
-                elif node.func.id == 'set':
-                    yield (node.lineno, node.col_offset, self.message_C101, type(self))
-                elif node.func.id == 'dict':
-                    yield (node.lineno, node.col_offset, self.message_C102, type(self))
+                if node.func.id in ('list', 'set', 'dict'):
+                    yield (
+                        node.lineno,
+                        node.col_offset,
+                        self.message_C400.format(type=node.func.id),
+                        type(self),
+                    )